Move grpc config to plugin/storage
